+++ bubble-preload.js
@@ -2,6 +2,7 @@ const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('drippyBubble', {
   onData: (cb) => ipcRenderer.on('bubble:data', (_e, data) => cb(data)),
+  onTail: (cb) => ipcRenderer.on('bubble:tail', (_e, t) => cb(t)),
   hover: (over) => ipcRenderer.send('drippy:bubble-hover', { over }),
   action: () => ipcRenderer.send('drippy:bubble-action'),
   reportHeight: (h) => ipcRenderer.send('drippy:bubble-height', { h }),